core: tasks: settle-match: Prove `VALID MATCH ENCRYPTION` in task
