+++ core/src/handshake/mod.rs
@@ -1,5 +1,4 @@
 //! The handshake module handles performing MPC handshakes with peers
-mod encumber;
 pub mod error;
 mod handshake_cache;
 pub mod jobs;
